got messages working and users adding to main chat channel list
